add testing and enable delete profile button

+++ server/profilePicture.py
@@ -4,9 +4,9 @@ from mysql.connector import errorcode
 from server.connect import connectToDB
 
 def saveImage(image, userId, newUser = False):
+    if not newUser:
+        deletePreviousImage(userId)
     if(image.filename != ''):
-        if not newUser:
-            deletePreviousImage(userId)
         filename = uuid.uuid4().hex 
         extension = os.path.splitext(image.filename)[1]
         fullname= filename + extension
@@ -14,7 +14,7 @@ def saveImage(image, userId, newUser = False):
         return "profilePictures/"+fullname
     return None
 
-def deletePreviousImage(userId):
+def deletePreviousImage(userId, testing = False):
     try:
         connection = connectToDB()
         if(connection != False):
@@ -23,12 +23,10 @@ def deletePreviousImage(userId):
             cursor.execute(sql)
             result = cursor.fetchone()[0]
             connection.commit()
-            if result is not None:
+            if result is not None and not testing:
                 os.remove("../public/"+result)
-            return None
+            return "../public/"+result
     except mysql.connector.Error as err:
 
         return {"response": err.msg }
-
-    print("FUck")
